+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/dataresources/products/ProductsDbConnector.groovy
@@ -142,8 +142,8 @@ class ProductsDbConnector implements ArchiveProductDataSource, CreateProductData
       String query = "INSERT INTO productitem (productId, quantity, offerid) "+
               "VALUE(?,?,?)"
 
-      //todo here could occur a potential issue, what if the products cannot be found (were archived during the process of offer creation)
-      int productId = findProductId(productItem.product)
+      Integer productId = getProductPrimaryId(productItem.product)
+              .orElseThrow({new DatabaseQueryException("Could not determine product primary id for ${productItem.product}!")})
 
       provider.connect().withCloseable {
         PreparedStatement preparedStatement = it.prepareStatement(query)
@@ -154,6 +154,13 @@ class ProductsDbConnector implements ArchiveProductDataSource, CreateProductData
         preparedStatement.execute()
       }
     }
+  }
+
+  private Optional<Integer> getProductPrimaryId(Product product) {
+    def result = findProductId(product)
+    Optional<Integer> productPrimaryId = Optional.ofNullable(result).flatMap({ it as Integer})
+    return productPrimaryId
+    return productPrimaryId
   }
 
   /**