+++ CHANGELOG.rst
@@ -5,11 +5,12 @@ Changelog
 This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 1.2.1 (2021-10-27)
-------------------
 
 **Added**
 
 **Fixed**
+
+* Introduce proper exception handling when product cannot found in the database (DM-66)
 
 **Dependencies**